+++ gulpfile.js
@@ -9,12 +9,12 @@
  */
 'use strict';
 
+const concat = require('gulp-concat');
 const depcheck = require('depcheck');
 const fs = require('fs');
 const glob = require('glob');
 const gulp = require('gulp');
 const bower = require('gulp-bower');
-const jshint = require('gulp-jshint');
 const mocha = require('gulp-spawn-mocha');
 const tslint = require('gulp-tslint');
 const ts = require('gulp-typescript');
@@ -22,6 +22,7 @@ const lazypipe = require('lazypipe');
 const path = require('path');
 const rollup = require('rollup');
 const runSequence = require('run-sequence');
+const typescript = require('typescript');
 
 
 // const commonTools = require('tools-common/gulpfile');
@@ -29,12 +30,7 @@ const commonTools = {
   depcheck: commonDepCheck
 };
 
-const tsProject = ts.createProject('tsconfig.json', {
-  typescript: require('typescript')
-});
-
-
-gulp.task('lint', ['tslint', 'test:style', 'depcheck']);
+gulp.task('lint', ['tslint', 'depcheck']);
 
 // Meta tasks
 
@@ -57,22 +53,28 @@ function removeFile(path) {
 gulp.task('clean', (done) => {
   removeFile('browser.js');
   removeFile('browser.js.map');
-  glob('runner/*.js', (err, files) => {
-    if (err) return done(err);
-    try {
-      for (const file of files) {
-        removeFile(file);
+  const patterns = ['runner/*.js', 'browser/**/*.js', 'browser/**/*.js.map'];
+  for (const pattern of patterns) {
+    glob(pattern, (err, files) => {
+      if (err) {
+        return done(err);
       }
-    } catch (e) {
-      return done(e);
-    }
-    done();
-  });
+      try {
+        for (const file of files) {
+          removeFile(file);
+        }
+      } catch (e) {
+        return done(e);
+      }
+    });
+  }
+  done();
 });
 
 gulp.task('test', function (done) {
   runSequence(
-    ['build:typescript', 'lint'],
+    'build:typescript-server',
+    'lint',
     'test:unit',
     'test:integration',
     done);
@@ -82,17 +84,27 @@ gulp.task('build-all', (done) => {
   runSequence('clean', 'lint', 'build', done);
 });
 
-gulp.task('build', ['build:typescript', 'build:browser', 'build:wct-browser-legacy']);
+gulp.task('build',
+  ['build:typescript-server', 'build:browser', 'build:wct-browser-legacy']);
 
-gulp.task('build:typescript', function () {
+const tsProject = ts.createProject('tsconfig.json', { typescript });
+gulp.task('build:typescript-server', function () {
   // Ignore typescript errors, because gulp-typescript, like most things
   // gulp, can't be trusted.
   return tsProject.src().pipe(tsProject(ts.reporter.nullReporter())).js.pipe(gulp.dest('./'));
 });
 
+const browserTsProject = ts.createProject('browser/tsconfig.json', {
+  typescript
+});
+gulp.task('build:typescript-browser', function () {
+  return browserTsProject.src().pipe(
+    browserTsProject(ts.reporter.nullReporter())).js.pipe(gulp.dest('./browser/'));
+});
+
 // Specific tasks
 
-gulp.task('build:browser', function (done) {
+gulp.task('build:browser', ['build:typescript-browser'], function (done) {
   rollup.rollup({
     entry: 'browser/index.js',
   }).then(function (bundle) {
@@ -110,7 +122,13 @@ gulp.task('build:browser', function (done) {
   }).catch(done);
 });
 
-gulp.task('build:browser-for-wct-browser-legacy', function (done) {
+gulp.task('build:wct-browser-legacy:a11ySuite', function () {
+  return gulp.src(['data/a11ySuite-npm-header.txt', 'data/a11ySuite.js'])
+    .pipe(concat('a11ySuite.js'))
+    .pipe(gulp.dest('wct-browser-legacy/'));
+});
+
+gulp.task('build:wct-browser-legacy:browser', ['build:typescript-browser'], function (done) {
   rollup.rollup({
     entry: 'browser/index.js',
   }).then(function (bundle) {
@@ -128,18 +146,11 @@ gulp.task('build:browser-for-wct-browser-legacy', function (done) {
   }).catch(done);
 });
 
-gulp.task('build:wct-browser-legacy', ['build:browser-for-wct-browser-legacy'], function () {
-  return gulp.src(['data/a11ySuite.js'])
-    .pipe(gulp.dest('wct-browser-legacy/'));
-});
+gulp.task('build:wct-browser-legacy', [
+  'build:wct-browser-legacy:a11ySuite',
+  'build:wct-browser-legacy:browser',
+]);
 
-gulp.task('test:style', function () {
-  return gulp.src([
-    '{browser,runner,environment,tasks}/**/*.js',
-    'gulpfile.js',
-    '!runner/*.js',
-  ]).pipe(jshintFlow());
-});
 
 gulp.task('test:unit', function () {
   return gulp.src('test/unit/*.js', { read: false })
@@ -159,19 +170,12 @@ gulp.task('tslint', () =>
   gulp.src([
     'runner/**/*.ts', '!runner/**/*.d.ts',
     'test/**/*.ts', '!test/**/*.d.ts',
-    'custom_typings/*.d.ts'
+    'custom_typings/*.d.ts', 'browser/**/*.ts', '!browser/**/*.ts'
   ])
-    .pipe(tslint({
-      configuration: 'tslint.json',
-    }))
-    .pipe(tslint.report('verbose')));
+    .pipe(tslint())
+    .pipe(tslint.report({ formatter: 'verbose' })));
 
 // Flows
-
-const jshintFlow = lazypipe()
-  .pipe(jshint)
-  .pipe(jshint.reporter, 'jshint-stylish')
-  .pipe(jshint.reporter, 'fail');
 
 commonTools.depcheck({
   stickyDeps: new Set([