+++ runner/test.ts
@@ -81,10 +81,10 @@ export async function test(options: Config|Context): Promise<void> {
     await steps.runTests(context);
   } finally {
     if (!context.options.skipCleanup) {
-      await new Promise((resolve) => cleankill.close(resolve));
+      await cleankill.close();
     }
   }
-};
+}
 
 // HACK
 test['test'] = test;