+++ test/unit/context.ts
@@ -36,7 +36,7 @@ describe('Context', () => {
 
     it('excludes plugins with a falsy config', async () => {
       const context = new Context(<any>{plugins: {local: false, sauce: {}}});
-      const stub = sandbox.stub(Plugin, 'get', (name: string) => {
+      const stub = sandbox.stub(Plugin, 'get').callsFake((name: string) => {
         return Promise.resolve(name);
       });
 
@@ -49,7 +49,7 @@ describe('Context', () => {
     it('excludes plugins disabled: true', async () => {
       const context =
           new Context(<any>{plugins: {local: {}, sauce: {disabled: true}}});
-      const stub = sandbox.stub(Plugin, 'get', (name: string) => {
+      const stub = sandbox.stub(Plugin, 'get').callsFake((name: string) => {
         return Promise.resolve(name);
       });